backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.16
python-jose[cryptography]==3.3.0
psycopg2-binary
python-dotenv==1.0.0
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0